pyproject.toml
[project]
name = "surgical-robot-tremor-compensator-rl"
version = "0.1.0"
description = "RL agent (SAC) for real-time surgical robot tremor compensation"
requires-python = ">=3.11"
license = "Apache-2.0"
dependencies = [
    "gymnasium>=0.29.0",
    "stable-baselines3>=2.1.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pyyaml>=6.0",
    "torch>=2.0.0",
    "scipy>=1.10.0",
    "mlflow>=2.10.0",
    "prometheus-client>=0.20.0",
    "tensorboard>=2.20.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-benchmark>=4.0.0",
    "ruff>=0.4.0",
    "py-spy>=0.3.0",
    "snakeviz>=2.2.0",
    "pip-audit>=2.6.0",
    "cyclonedx-bom>=4.0.0",
    "httpx>=0.27.0",
]
physics = [
    "pybullet>=3.2.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["env", "agents", "safety", "dashboard", "utils"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "benchmark: marks tests as benchmark tests",
]
